Try not to delete updated AppImages any more, mk. 1

+++ src/ui/update_main.cpp
@@ -193,24 +193,32 @@ int main(int argc, char** argv) {
         return 1;
     }
 
-    if (!appimage_shall_not_be_integrated(pathToAppImage.toStdString().c_str())) {
-        const auto pathToIntegratedAppImage = buildPathToIntegratedAppImage(pathToAppImage);
+    const auto pathToIntegratedAppImage = buildPathToIntegratedAppImage(pathToAppImage);
 
+    if (!appimage_shall_not_be_integrated(pathToAppImage.toStdString().c_str())) {
         if (!integrateAppImage(pathToUpdatedAppImage, pathToIntegratedAppImage)) {
             criticalUpdaterError(QObject::tr("Failed to register updated AppImage in system"));
             return 1;
         }
     }
 
-    if (removeAfterUpdate) {
-        if (hasBeenRegisteredBefore && appimage_unregister_in_system(pathToAppImage.toStdString().c_str(), false) != 0) {
-            criticalUpdaterError(QObject::tr("Failed to unregister old AppImage in system"));
-            return 1;
-        }
+    // a crappy attempt to prevent deletion of the updated AppImage in a rare case (see below)
+    const auto pathToIntegratedUpdatedAppImage = buildPathToIntegratedAppImage(pathToUpdatedAppImage);
 
-        if (!QFile::remove(pathToAppImage)) {
-            criticalUpdaterError(QObject::tr("Failed to remove old AppImage"));
-            return 1;
+    if (removeAfterUpdate) {
+        // make sure not to delete the updated(!) AppImage if the filenames of the new and old file are equal
+        // in this case, a warning is shown, asking the user whether to overwrite the old file, and in that case we
+        // don't need to unregister nor delete the file
+        if (pathToIntegratedAppImage != pathToIntegratedUpdatedAppImage) {
+            if (hasBeenRegisteredBefore && appimage_unregister_in_system(pathToAppImage.toStdString().c_str(), false) != 0) {
+                criticalUpdaterError(QObject::tr("Failed to unregister old AppImage in system"));
+                return 1;
+            }
+
+            if (!QFile::remove(pathToAppImage)) {
+                criticalUpdaterError(QObject::tr("Failed to remove old AppImage"));
+                return 1;
+            }
         }
     }
